Move telemetry to extension
Move the OpenTelemetry code into (mostly) one place separate from the rest of the function code.
